Move Figure.shift_origin tests into a separate test file
